+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to `async` will be documented in this file
 ## 1.5.2 - 2020-11-20
 
 - Configure task in synchronous process
+- Add Pool::forceSynchronous function
 
 ## 1.5.1 - 2020-11-20